+++ pallets/carbon-credits/Cargo.toml
@@ -21,7 +21,7 @@ sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "po
 bs58 = { default-features = false, version = '0.4.0',features=['alloc']}
 scale-info = { version = "2.0.1", default-features = false, features = ["derive"] }
 pallet-assets = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.29", default-features = false }
-primitives = { package = 'bitg-primitives', path = '../../primitives', default-features = false }
+primitives = { package = 'bitgreen-primitives', path = '../../primitives', default-features = false }
 pallet-balances = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.29", default-features = false }
 pallet-uniques = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.29", default-features = false }
