+++ parachain/src/command.rs
@@ -27,10 +27,9 @@ use crate::{
 fn load_spec(id: &str) -> std::result::Result<Box<dyn ChainSpec>, String> {
 	Ok(match id {
 		"dev" => Box::new(chain_spec::development_config()),
-		"template-rococo" => Box::new(chain_spec::local_testnet_config()),
 		"rococo" => Box::new(chain_spec::rococo_config()),
 		"" | "local" => Box::new(chain_spec::local_testnet_config()),
-		path => Box::new(chain_spec::ChainSpec::from_json_file(
+		path => Box::new(chain_spec::RococoChainSpec::from_json_file(
 			std::path::PathBuf::from(path),
 		)?),
 	})
