+++ primitives/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "bitg-primitives"
+name = "bitgreen-primitives"
 version = "0.7.3"
 authors = ['BitGreen <https://www.bitgreen.org>']
 edition = "2021"
@@ -7,15 +7,11 @@ edition = "2021"
 [dependencies]
 serde = { version = "1.0.136", optional = true }
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["max-encoded-len"] }
-parity-util-mem = { version = "0.11", default-features = false, features = ["primitive-types"] }
 scale-info = { version = "2.0.1", default-features = false, features = ["derive"] }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.29", default-features = false }
+frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.29", default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.29", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.29", default-features = false }
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.29", default-features = false }
-
-[dev-dependencies]
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.29" }
 
 [features]
 default = ["std"]
@@ -25,5 +21,6 @@ std = [
 	"sp-runtime/std",
 	"sp-core/std",
 	"sp-std/std",
-	"scale-info/std"
+	"scale-info/std",
+	"frame-support/std"
 ]
