--- runtime/rococo/src/benchmarking/mod.rs
@@ -1,9 +1,0 @@
-#![cfg(feature = "runtime-benchmarks")]
-
-// module benchmarking
-pub mod evm;
-pub mod evm_accounts;
-
-// orml benchmarking
-pub mod authority;
-pub mod utils;
